fix(cache): correct the scope to be cached
It should be window-locally to check if this function has been executed
to a commit buffer; a buffer to edit commit message could possibly be
reused at several commit hashes up to users.

On the other hand, the window would be rarely reused.
The reason is because either aborting or updating a commit message, the
window will be closed with the commit buffer.
Addition to the above, Vim will open a new window when we start another
commit message which belongs to another git-repository, while editing
a commit message.